+++ src/migrations/m180531_090631_v2_upgrade.php
@@ -50,7 +50,7 @@ class m180531_090631_v2_upgrade extends Migration
         $tablesCreated = false;
 
         // support_tickets table
-        $tableSchema = Craft::$app->db->schema->getTableSchema('{{%support_tickets}}');
+        $tableSchema = Craft::$app->db->schema->getTableSchema('{{%checkout_charges}}');
         if ($tableSchema === null) {
             $tablesCreated = true;