+++ CHANGELOG.md
@@ -18,3 +18,4 @@ All notable changes to this project will be documented in this file.
 - Improved settings templates
 - Event names changed to `EVENT_BEFORE_CHARGE` and `EVENT_AFTER_CHARGE`
 - Updated icons
+- More charge data made available