+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.0.1 - 2018-05-31
+
+### Fixed
+
+- Migration table schema
+
 ## 2.0.0 - 2018-05-31
 
 ### Added